// src/types/VocabularyBeta.ts
export interface VocabularyBeta {
  vocabId: string;
  wordType: string;
  germanWord: string;
  englishTranslation: string;
  pronunciationIpa: string;
  exampleDe: string;
  exampleEn: string;
  category: string;
  tags: string;
  frequency: number;
  level: number;
  difficulty: number;
  verbStem: string;
  verbType: string;
  pastParticiple: string;
  comparative: string;
  superlative: string;
  governsCase: string;
  usageNotes: string;
  culturalNote: string;
  id: string;
}